+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ---
 ## [1.0.6] - Unreleased
+### Improved
+- Optimized Props of the Quarantine Table + A Loader tweak
 
 ## [1.0.5] - 2026-05-03
 ### Added